+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
 ## [3.2.0] – 2019-09-01
 ### Added
 - In `SchemaGenerator.generateSchema(Type)` also allow passing type parameters via `SchemaGenerator.generateSchema(Type, Type...)`.
@@ -66,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Specific handling of optionals (two alternatives via standard options).
 - Pre-defined sets of standard options to cover different use-cases and simplify library usage.
 
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v3.2.0...HEAD
 [3.2.0]: https://github.com/victools/jsonschema-generator/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/victools/jsonschema-generator/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/victools/jsonschema-generator/compare/v2.0.0...v3.0.0