+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.3.0] - 2019-10-25
 ### Fixed
 - Increase dependency version for jackson-databind (and jackson-core) to resolve security alerts.
-- Avoid unnecessary quotes when representing constant string values (due to changed jackson behaviour)
+- Avoid unnecessary quotes when representing constant string values (due to changed behaviour in jackson >=2.10.0)
 
 ## [3.2.0] – 2019-09-01
 ### Added
@@ -71,7 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Specific handling of optionals (two alternatives via standard options).
 - Pre-defined sets of standard options to cover different use-cases and simplify library usage.
 
-[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v3.2.0...HEAD
+[3.3.0]: https://github.com/victools/jsonschema-generator/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/victools/jsonschema-generator/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/victools/jsonschema-generator/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/victools/jsonschema-generator/compare/v2.0.0...v3.0.0